Refactor: Split GitDialogs.tsx into separate files
Split the 1426-line GitDialogs.tsx monolith into smaller, focused modules:
- git/shared.tsx: Types, BaseDialog, BranchSelector, and shared UI components
- git/MergeDialog.tsx: Merge branch dialog
- git/RebaseDialog.tsx: Rebase branch dialog
- git/PRDialog.tsx: Create pull request dialog
- git/SquashDialog.tsx: Squash commits dialog
- git/ForcePushDialog.tsx: Force push confirmation dialog
- git/useGitDialogs.ts: Main hook with all git operation logic
- git/index.ts: Re-exports for backwards compatibility

This improves maintainability by:
- Reducing file size from 1426 lines to ~100-300 lines per file
- Isolating each dialog's concerns
- Making the codebase easier to navigate and modify
